.requirements.txt
pytest
jsonschema
referencing
ipython
pyyaml
ga4gh.gks.metaschema==0.3.0b4
sphinx ~= 7.2
sphinx-rtd-theme ~= 1.2